+++ CHANGELOG.md
@@ -1,3 +1,47 @@
+## [1.0.0] - 2025-01-02
+### Added
+- Organized files
+- Code documentation
+
+## [0.7.0] - 2025-01-01
+### Added
+- Added trump card gui input
+
+## [0.6.0] - 2024-12-26
+### Added
+- More Game UI development
+
+## [0.5.0] - 2024-12-25
+### Added
+- Major Game UI development
+
+### Changed
+- Revised "game.py"
+
+## [0.4.0] - 2024-12-07
+### Added
+- Main Menu UI development
+
+### Changed
+- Removed module "main_menu.py".
+- Updated "viewStats" method inside "userstats.py" module.
+
+## [0.3.0] - 2024-12-01
+### Added
+- Main Menu UI development
+
+### Changed
+- Updated README.md
+- Decomposed 'enterProfile' method into smaller methods
+- New folder 'gui' contains methods using tkinter 
+
+## [0.2.0] - 2024-11-24
+### Added
+- Login page UI development
+
+### Fixed
+- Fixed being able to create a new profile with empty username/password
+
 ## [0.1.0] - 2024-11-20
 ### Added
 - Initial version of (currently) terminal based blackjack game